ai-ml-gcp/requirements.txt
<<<<<<< HEAD
# ── Core pipeline ──────────────────────────────────────────────────────────────
pandas>=2.0
openpyxl>=3.1
rapidfuzz>=3.5
typer>=0.12
pydantic>=2.6
pyyaml>=6.0
rich>=13.7
requests>=2.31
torch>=2.0.0
torchvision>=0.16.0

# ── YouTube / transcript ───────────────────────────────────────────────────────
yt-dlp>=2024.7.0
webvtt-py>=0.5
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0

# ── Embeddings (local, no data leaves machine) ────────────────────────────────
sentence-transformers>=2.2
scikit-learn>=1.3
numpy>=1.24

# ── Qdrant vector store (local Docker or in-memory) ───────────────────────────
qdrant-client>=1.9.0

# ── Voice pipeline ─────────────────────────────────────────────────────────────
edge-tts>=6.1.9
imageio-ffmpeg>=0.6.0

# ── LiveKit (voice pipeline) ──────────────────────────────────────────────────
livekit>=0.11
livekit-agents>=0.8

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.31.0

# ── API server ───────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0
gunicorn>=22.0.0        
python-multipart>=0.0.9   

# ── Optional cloud STT (uncomment if using Deepgram) ──────────────────────────
# deepgram-sdk>=3.0
=======
# ── Core pipeline ──────────────────────────────────────────────────────────────
pandas>=2.0
openpyxl>=3.1
rapidfuzz>=3.5
typer>=0.12
pydantic>=2.6
pyyaml>=6.0
rich>=13.7
requests>=2.31
torch>=2.0.0
torchvision>=0.16.0

# ── YouTube / transcript ───────────────────────────────────────────────────────
yt-dlp>=2024.7.0
webvtt-py>=0.5
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0

# ── Embeddings (local, no data leaves machine) ────────────────────────────────
sentence-transformers>=2.2
scikit-learn>=1.3
numpy>=1.24

google-generativeai>=0.8.0
pymongo[srv]>=4.7.0


# ── Qdrant vector store (local Docker or in-memory) ───────────────────────────
qdrant-client>=1.9.0

# ── Voice pipeline ─────────────────────────────────────────────────────────────
edge-tts>=6.1.9
pydub>=0.25.1

# ── LiveKit (voice pipeline) ──────────────────────────────────────────────────
livekit>=0.11
livekit-agents>=0.8

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.31.0

# ── API server ───────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0
gunicorn>=22.0.0        
python-multipart>=0.0.9   

# ── Optional cloud STT (uncomment if using Deepgram) ──────────────────────────
# deepgram-sdk>=3.0
>>>>>>> 8f5668b7ef0ae7ff95342c07a5e69a2922229415
# sounddevice   # for souli voice --local (mic/speaker test)